requirements.txt
streamlit
pandas
numpy
scikit-learn
openai
sentence-transformers
torch
python-dotenv
matplotlib  
ipykernel
plotly